refactor: move mutation and rootQuery out of the main file && reduce method duplication
